requirements.txt
# Core ML/DL frameworks
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
deepspeed>=0.12.0
bitsandbytes>=0.41.0

# Mamba architecture
mamba-ssm>=1.2.0
causal-conv1d>=1.0.0

# Data processing
datasets>=2.14.0
tokenizers>=0.15.0
huggingface-hub>=0.19.0
webdataset>=0.2.0

# Training utilities
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0

# Distributed training
mpi4py>=3.1.0
nccl>=2.18.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Utilities
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
psutil>=5.9.0
